convert legacy LWRP to custom resources
