+++ resources/install_global.rb
@@ -5,15 +5,58 @@
 # Copyright:: 2012-2021, Escape Studios
 #
 
-actions :install, :update, :remove
 default_action :install
 
-attribute :package, kind_of: String, name_attribute: true, required: true
-attribute :version, kind_of: String, default: '*.*.*'
-attribute :install_dir, kind_of: String, default: nil
-attribute :bin_dir, kind_of: String, default: nil
-attribute :dev, kind_of: [TrueClass, FalseClass], default: false
-attribute :quiet, kind_of: [TrueClass, FalseClass], default: true
-attribute :optimize_autoloader, kind_of: [TrueClass, FalseClass], default: false
-attribute :prefer_dist, kind_of: [TrueClass, FalseClass], default: false
-attribute :prefer_source, kind_of: [TrueClass, FalseClass], default: false
+property :package, String, name_property: true
+property :version, String, default: '*.*.*'
+property :install_dir, String, default: lazy { node['composer']['global_install']['install_dir'] }
+property :bin_dir, String, default: lazy { node['composer']['global_install']['bin_dir'] }
+property :dev, [true, false], default: false
+property :quiet, [true, false], default: true
+property :optimize_autoloader, [true, false], default: false
+property :prefer_dist, [true, false], default: false
+property :prefer_source, [true, false], default: false
+
+action :install do
+  directory new_resource.install_dir
+
+  composer_project new_resource.install_dir do
+    package new_resource.package
+    version new_resource.version
+    bin_dir new_resource.bin_dir
+    dev new_resource.dev
+    quiet new_resource.quiet
+    optimize_autoloader new_resource.optimize_autoloader
+    prefer_dist new_resource.prefer_dist
+    prefer_source new_resource.prefer_source
+    action :require
+  end
+end
+
+action :remove do
+  composer_project new_resource.install_dir do
+    package new_resource.package
+    version new_resource.version
+    bin_dir new_resource.bin_dir
+    dev new_resource.dev
+    quiet new_resource.quiet
+    optimize_autoloader new_resource.optimize_autoloader
+    prefer_dist new_resource.prefer_dist
+    prefer_source new_resource.prefer_source
+    action :remove
+  end
+end
+
+action :update do
+  composer_project new_resource.install_dir do
+    package new_resource.package
+    version new_resource.version
+    bin_dir new_resource.bin_dir
+    dev new_resource.dev
+    quiet new_resource.quiet
+    optimize_autoloader new_resource.optimize_autoloader
+    prefer_dist new_resource.prefer_dist
+    prefer_source new_resource.prefer_source
+    action :update
+  end
+end
